requirements.txt
crawl4ai>=0.7.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354
spacy>=3.6.0
fast-langdetect>=1.0.0
transformers>=4.35.0
torch>=2.0.0
scikit-learn>=1.3.0
keybert>=0.8.3

# GUI
dearpygui>=1.11.0

# PDF Report Generation
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
reportlab>=4.0.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
pypdf>=4.0.0

# CLI Framework
typer>=0.9.0
rich>=13.0.0
